step_05/testbedCPO/requirements.txt
fastapi
uvicorn
pydantic
openleadr
websockets
pandas
ocpp
python-dotenv
